Move IL offset projection into Research
Establish top-level projection contracts, a focused ILOffsetProjectionProducer, and a thin ResearchViews forwarder. Reduce the CLI query to parsing, capability selection, symbol acquisition, and failure handling while preserving output and keeping PE/instruction composition below the tool layer.


Copilot-Session: b7300045-6e27-433f-9dfc-f51c4760452e
