+++ src/ILInspector.Metadata/ILInspector.Metadata.csproj
@@ -16,6 +16,7 @@
   <ItemGroup>
     <InternalsVisibleTo Include="dotnet-inspect" />
     <InternalsVisibleTo Include="dotnet-inspect.Tests" />
+    <InternalsVisibleTo Include="ILInspector.Research" />
     <InternalsVisibleTo Include="ILInspector.Metadata.Tests" />
   </ItemGroup>
