+++ src/ILInspector.Research/ResearchViews.ILOffset.cs
@@ -1,0 +1,7 @@
+namespace ILInspector.Research;
+
+public static partial class ResearchViews
+{
+    public static ILOffsetProjectionOutcome ProjectILOffset(ILOffsetProjectionRequest request)
+        => ILOffsetProjectionProducer.Produce(request);
+}
