+++ src/dotnet-inspect/Models/LibraryInspection.cs
@@ -1,8 +1,9 @@
 using System.Text.Json.Serialization;
 using DotnetInspector.Options;
+using ILInspector.Analysis;
 using ILInspector.Findings;
 using ILInspector.Metadata;
-using ILInspector.Analysis;
+using ILInspector.Research;
 
 namespace DotnetInspector.Models;
 
@@ -686,7 +687,7 @@ public class LibraryInspection
     /// an IL coordinate section is selected with a token+offset parameter.
     /// </summary>
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public ILOffsetResult? ILOffset { get; set; }
+    public ILOffsetProjection? ILOffset { get; set; }
 
     // Presence flags — populated cheaply from MetadataReader before scanners run.
     // Used by CanRender for fast -s discovery without full scanning.
@@ -912,136 +913,6 @@ public class LibraryInspection
     /// </summary>
     [JsonIgnore]
     public bool UseDependenciesView { get; set; }
-}
-
-public class ILOffsetResult
-{
-    public string? Method { get; init; }
-    public string? Token { get; init; }
-    public string? ILOffset { get; init; }
-    public string? MatchedOffset { get; init; }
-    public string? File { get; init; }
-    public int? Line { get; init; }
-    public string? Url { get; init; }
-    public ILOffsetMemberContext? MemberContext { get; init; }
-    public ILOffsetInstructionContext? InstructionContext { get; init; }
-    public List<ILOffsetExceptionContext>? ExceptionContext { get; init; }
-    public ILOffsetCallsiteContext? CallsiteContext { get; init; }
-    public ILOffsetReturnAddressContext? ReturnAddressContext { get; init; }
-    public List<ILOffsetAllocationContext>? AllocationContext { get; init; }
-    public List<ILOffsetSafetyContext>? SafetyContext { get; init; }
-    public List<ILOffsetCostContext>? CostContext { get; init; }
-}
-
-public class ILOffsetMemberContext
-{
-    public string? Assembly { get; init; }
-    public string? Type { get; init; }
-    public string? TypeKind { get; init; }
-    public string? Member { get; init; }
-    public string? Signature { get; init; }
-    public string? MemberKind { get; init; }
-    public string? Visibility { get; init; }
-    public string? Static { get; init; }
-    public string? Async { get; init; }
-    public string? MetadataToken { get; init; }
-    public string? ILOffset { get; init; }
-}
-
-public class ILOffsetInstructionContext
-{
-    public string? ILOffset { get; init; }
-    public string? Boundary { get; init; }
-    public string? Opcode { get; init; }
-    public string? OperandKind { get; init; }
-    public string? Operand { get; init; }
-    public string? OperandToken { get; init; }
-    public string? BranchTargets { get; init; }
-    public string? NextOffset { get; init; }
-    public int? Length { get; init; }
-    public int? Block { get; init; }
-    public string? TerminatesBlock { get; init; }
-    public string? FallsThrough { get; init; }
-}
-
-public class ILOffsetExceptionContext
-{
-    public int Region { get; init; }
-    public string? Context { get; init; }
-    public string? Clause { get; init; }
-    public string? TryRange { get; init; }
-    public string? HandlerRange { get; init; }
-    public string? FilterRange { get; init; }
-    public string? CaughtType { get; init; }
-}
-
-public class ILOffsetCallsiteContext
-{
-    public string? CallOffset { get; init; }
-    public string? Opcode { get; init; }
-    public string? CallKind { get; init; }
-    public string? Callee { get; init; }
-    public string? OperandToken { get; init; }
-    public string? ReturnAddress { get; init; }
-}
-
-public class ILOffsetReturnAddressContext
-{
-    public string? ILOffset { get; init; }
-    public string? CallOffset { get; init; }
-    public string? Opcode { get; init; }
-    public string? CallKind { get; init; }
-    public string? Callee { get; init; }
-    public string? OperandToken { get; init; }
-}
-
-public class ILOffsetAllocationContext
-{
-    public string? ILOffset { get; init; }
-    public string? AllocationKind { get; init; }
-    public string? AllocatedType { get; init; }
-    public string? CountedAsHeap { get; init; }
-    public string? Frequency { get; init; }
-    public string? Escape { get; init; }
-    [JsonPropertyName("escape_kind")]
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public string? EscapeKind { get; init; }
-    [JsonPropertyName("est_size")]
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public int? EstimatedSizeBytes { get; init; }
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public string? SizeTier { get; init; }
-    public string? InLoop { get; init; }
-    public string? Path { get; init; }
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public string? PathConfidence { get; init; }
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public string? PostDominance { get; init; }
-    public string? Evidence { get; init; }
-    [JsonPropertyName("multiplicity")]
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public string? Multiplicity { get; init; }
-    [JsonPropertyName("churned_type")]
-    [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public string? ChurnedType { get; init; }
-}
-
-public class ILOffsetSafetyContext
-{
-    public string? ILOffset { get; init; }
-    public string? SafetyKind { get; init; }
-    public string? Operation { get; init; }
-    public string? Requirement { get; init; }
-    public string? Evidence { get; init; }
-}
-
-public class ILOffsetCostContext
-{
-    public string? ILOffset { get; init; }
-    public string? CostKind { get; init; }
-    public string? Operation { get; init; }
-    public string? InLoop { get; init; }
-    public string? Evidence { get; init; }
 }
 
 public sealed record SourceFileInfo(string Type, string? Url);
